% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine.R, R/combine_result.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,n2kAnalysisVersion-method}
\alias{combine,n2kAnalysisVersion-methods}
\alias{combine,n2kAnamoly-methods}
\alias{combine,n2kAnomaly-method}
\alias{combine,n2kParameter-method}
\alias{combine,n2kParameter-methods}
\alias{combine,n2kResult-method}
\alias{combine,n2kResult-methods}
\title{Combine multiple n2k objects}
\usage{
combine(...)

\S4method{combine}{n2kAnalysisVersion}(...)

\S4method{combine}{n2kParameter}(...)
}
\arguments{
\item{...}{n2k objects}
}
\description{
Combine multiple n2k objects
}

