% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_s3.R
\name{connect_inbo_s3}
\alias{connect_inbo_s3}
\title{Set environment variables for INBO S3 bucket}
\usage{
connect_inbo_s3()
}
\description{
Before running this function you must have an \code{.aws} folder in your home
directory with a \code{credentials} file containing the credentials for the INBO
shared infrastructure.
Run the \verb{aws assume role} command to get the credentials for the INBO shared
infrastructure before running this function.
}
