% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_model.R
\docType{methods}
\name{delete_model}
\alias{delete_model}
\alias{delete_model,ANY,character-method}
\alias{delete_model,ANY,s3_bucket-method}
\title{Delete an \code{n2kModel} object}
\usage{
delete_model(x, base, project)

\S4method{delete_model}{ANY,character}(x, base, project)

\S4method{delete_model}{ANY,s3_bucket}(x, base, project)
}
\arguments{
\item{x}{the file fingerprint of the \code{n2kModel}}

\item{base}{the base location}

\item{project}{will be a relative path within the base location}
}
\description{
Delete an \code{n2kModel} object
}
