% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display a message.}
\usage{
display(verbose, message, linefeed = TRUE)
}
\arguments{
\item{verbose}{A logical.
When \code{TRUE} print the message.
When \code{FALSE} do nothing.}

\item{message}{a vector passed to \code{\link[=message]{message()}}.}

\item{linefeed}{A logical.
When \code{TRUE} append a newline character at the end of the message.}
}
\description{
This is a short cut for \code{if(verbose) message(x)}.
}
