% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract,ANY,character-method}
\alias{extract,character-methods}
\alias{extract,ANY,n2kInlaNbinomial-method}
\alias{extract,n2kInlaNbinomial-methods}
\alias{extract,ANY,n2kInlaPoisson-method}
\alias{extract,n2kInlaPoisson-methods}
\title{Extract the relevant coefficients}
\usage{
extract(extractor, object, base, project)

\S4method{extract}{ANY,character}(extractor, object, base, project)

\S4method{extract}{ANY,n2kInlaNbinomial}(extractor, object, base = NULL,
  project = NULL)

\S4method{extract}{ANY,n2kInlaPoisson}(extractor, object, base = NULL,
  project = NULL)
}
\arguments{
\item{extractor}{the extractor function}

\item{object}{the n2kModel object}

\item{base}{the optional base location of the object}

\item{project}{the optional subdirectory}
}
\value{
the relevant coefficients
}
\description{
Extract the relevant coefficients
}
