% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_every_model.R
\name{fit_every_model}
\alias{fit_every_model}
\title{Fit the model to the analysis files}
\usage{
fit_every_model(path = ".", status = c("new", "error", "converged",
  "false convergence"))
}
\arguments{
\item{path}{the path containing the analysis files}

\item{status}{A vector of statusses for which the model will be calculated}
}
\description{
Fit the model to the analysis files
}

