% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_glmer_poisson.R
\name{fit_glmer_poisson}
\alias{fit_glmer_poisson}
\title{Fit a Poisson GLMM model  with lme4}
\usage{
fit_glmer_poisson(model, data, weight)
}
\arguments{
\item{model}{A character with the covariates of the model}

\item{data}{The data.frame with the data}

\item{weight}{Name of the weights variable}
}
\description{
Fit a Poisson GLMM model  with lme4
}

