% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_inla_nbinomial.R
\name{fit_inla_nbinomial}
\alias{fit_inla_nbinomial}
\title{Fit an INLA model with negative binomial distribution}
\usage{
fit_inla_nbinomial(model.set, data)
}
\arguments{
\item{model.set}{A character with the covariates of the model}

\item{data}{The data.frame with the data}
}
\value{
Either an INLA model or \code{"error"} in case the fitting yields an error
}
\description{
The function requires the INLA package
}

