% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\docType{methods}
\name{fit_model}
\alias{fit_model}
\alias{fit_model,character-method}
\alias{fit_model,n2kGlmerPoisson-method}
\alias{fit_model,n2kInlaNbinomial-method}
\title{Fit a n2kModel object}
\usage{
fit_model(x)

\S4method{fit_model}{character}(x)

\S4method{fit_model}{n2kGlmerPoisson}(x)

\S4method{fit_model}{n2kInlaNbinomial}(x)
}
\arguments{
\item{x}{the n2kModel}
}
\description{
Fit a n2kModel object
}

