% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_single_model.R
\name{fit_single_model}
\alias{fit_single_model}
\title{Run a single element from the to do list}
\usage{
fit_single_model(file, path)
}
\arguments{
\item{file}{The name of the rda file}

\item{path}{The path of the analysis files}
}
\description{
Run a single element from the to do list
}

