% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_analysis_date.R
\docType{methods}
\name{get_analysis_date}
\alias{get_analysis_date}
\alias{get_analysis_date,n2kModel-method}
\alias{get_analysis_date,n2kModel-methods}
\title{Get the analysis date of a n2kModel object}
\usage{
get_analysis_date(x)

\S4method{get_analysis_date}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the analysis date of the object
}
\description{
Get the analysis date of a n2kModel object
}

