% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_anomaly.R
\docType{methods}
\name{get_anomaly}
\alias{get_anomaly}
\alias{get_anomaly,n2kComposite-method}
\alias{get_anomaly,n2kComposite-methods}
\alias{get_anomaly,n2kGlmerPoisson-method}
\alias{get_anomaly,n2kGlmerPoisson-methods}
\alias{get_anomaly,n2kLrtGlmer-method}
\alias{get_anomaly,n2kLtrGlmer-methods}
\title{Get the anomalies from a model}
\usage{
get_anomaly(analysis, ...)

\S4method{get_anomaly}{n2kGlmerPoisson}(analysis, ...)

\S4method{get_anomaly}{n2kLrtGlmer}(analysis, ...)

\S4method{get_anomaly}{n2kComposite}(analysis, ...)
}
\arguments{
\item{analysis}{The model to add}

\item{...}{Extra options. See details}
}
\description{
Get the anomalies from a model
}

