% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_covariate.R
\docType{methods}
\name{get_covariate}
\alias{get_covariate}
\alias{get_covariate,n2kModel-method}
\alias{get_covariate,n2kModel-methods}
\title{Get the covariate of a n2kModel object}
\usage{
get_covariate(x)

\S4method{get_covariate}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the covariate of the object
}
\description{
Get the covariate of a n2kModel object
}

