% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data_fingerprint.R
\docType{methods}
\name{get_data_fingerprint}
\alias{get_data_fingerprint}
\alias{get_data_fingerprint,n2kModel-method}
\alias{get_data_fingerprint,n2kModel-methods}
\title{Get the data fingerprint of a n2kModel object}
\usage{
get_data_fingerprint(x)

\S4method{get_data_fingerprint}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the data fingerprint of the object
}
\description{
Get the data fingerprint of a n2kModel object
}

