% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datafield_id.R
\name{get_datafield_id}
\alias{get_datafield_id}
\title{Get the datafield id}
\usage{
get_datafield_id(table, field, datasource, root, stage = FALSE)
}
\arguments{
\item{table}{The table name}

\item{field}{The field name}

\item{datasource}{The datasource name}

\item{root}{The root of a project. Can be a file path or a \code{git-repository}.
Defaults to the current working directory (\code{"."}).}

\item{stage}{Logical value indicating whether to stage the changes after
writing the data. Defaults to \code{FALSE}.}
}
\description{
Get the datafield id
}
