% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_file_fingerprint.R
\docType{methods}
\name{get_file_fingerprint}
\alias{get_file_fingerprint}
\alias{get_file_fingerprint,n2kModel-method}
\alias{get_file_fingerprint,n2kModel-methods}
\title{Get the file fingerprint of a n2kModel object}
\usage{
get_file_fingerprint(x)

\S4method{get_file_fingerprint}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the file fingerprint of the object
}
\description{
Get the file fingerprint of a n2kModel object
}

