% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_formula.R
\docType{methods}
\name{get_formula}
\alias{get_formula}
\alias{get_formula,n2kAnalysisMetadata-method}
\alias{get_formula,n2kAnalysisMetadata-methods}
\title{Get the formula of a n2kModel object}
\usage{
get_formula(x)

\S4method{get_formula}{n2kAnalysisMetadata}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the formula of the object
}
\description{
Get the formula of a n2kModel object
}

