% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\docType{methods}
\name{get_model}
\alias{get_model}
\alias{get_model,n2kInla-method}
\alias{get_model,n2kInla-methods}
\alias{get_model,character-method}
\alias{get_model,character-methods}
\title{Get the model of a \code{n2kModel} object}
\usage{
get_model(x)

\S4method{get_model}{n2kInla}(x)

\S4method{get_model}{character}(x)
}
\arguments{
\item{x}{the \code{n2kModel} object}
}
\value{
the model of the object
}
\description{
Get the model of a \code{n2kModel} object
}
