% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model_set.R
\docType{methods}
\name{get_model_set}
\alias{get_model_set}
\alias{get_model_set,n2kModel-method}
\alias{get_model_set,n2kModel-methods}
\title{Get the model set of a n2kModel object}
\usage{
get_model_set(x)

\S4method{get_model_set}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the model set of the object
}
\description{
Get the model set of a n2kModel object
}

