% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_type.R
\docType{methods}
\name{get_model_type}
\alias{get_model_type}
\alias{get_model_type,n2kAnalysisMetadata-method}
\alias{get_model_type,n2kAnalysisMetadata-methods}
\title{Get the model type of an \code{n2kModel} object}
\usage{
get_model_type(x)

\S4method{get_model_type}{n2kAnalysisMetadata}(x)
}
\arguments{
\item{x}{the \code{n2kModel} object}
}
\value{
the model type of the object
}
\description{
Get the model type of an \code{n2kModel} object
}
