% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_parent.R
\docType{methods}
\name{get_parent}
\alias{get_parent}
\alias{get_parent,n2kLrtGlmer-method}
\alias{get_parent,n2kLrtGlmer-methods}
\title{Get the fingerprint of the parent analysis of a n2kModel object}
\usage{
get_parent(x, ...)

\S4method{get_parent}{n2kLrtGlmer}(x, ...)
}
\arguments{
\item{x}{the n2kModel object}

\item{...}{additional arguments (see details)}
}
\value{
the parent of the object
}
\description{
Get the fingerprint of the parent analysis of a n2kModel object
}
\details{
\describe{
 \item{\code{which.parent}}{0 = the parent for the null hypothesis, 1 =  the parent for the alternative hypothesis}
}
}

