% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parents.R
\name{get_parents}
\alias{get_parents}
\title{Get the object of the parent analyses of a n2kModel object}
\usage{
get_parents(child, base, project)
}
\arguments{
\item{child}{the child object}

\item{base}{the base location to read the model}

\item{project}{will be a relative path within the base location}
}
\description{
Get the object of the parent analyses of a n2kModel object
}
