% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R, R/get_result_character.R,
%   R/get_result_n2k_inla.R, R/get_result_n2k_model.R,
%   R/get_result_n2kmanifest.R, R/get_result_s3_object.R
\docType{methods}
\name{get_result}
\alias{get_result}
\alias{get_result,character,character-method}
\alias{get_result,character,s3_bucket-method}
\alias{get_result,n2kInla,ANY-method}
\alias{get_result,n2kModel,ANY-method}
\alias{get_result,n2kManifest,ANY-method}
\alias{get_result,s3_object,ANY-method}
\title{Add the results from an analysis}
\usage{
get_result(x, base, ...)

\S4method{get_result}{character,character}(x, base, ..., project, verbose = TRUE)

\S4method{get_result}{character,s3_bucket}(x, base, ..., project, verbose = TRUE)

\S4method{get_result}{n2kInla,ANY}(x, base, ..., verbose = TRUE)

\S4method{get_result}{n2kModel,ANY}(x, verbose = TRUE, ...)

\S4method{get_result}{n2kManifest,ANY}(x, base, ..., verbose = TRUE)

\S4method{get_result}{s3_object,ANY}(x, base, ..., project, verbose = TRUE)
}
\arguments{
\item{x}{object with the current results}

\item{base}{the base location to read the model}

\item{...}{further arguments (see Details)}

\item{project}{will be a relative path within the base location}

\item{verbose}{Print extra information on the screen}
}
\description{
Add the results from an analysis
}
