% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_scheme_id.R
\docType{methods}
\name{get_scheme_id}
\alias{get_scheme_id}
\alias{get_scheme_id,n2kModel-method}
\alias{get_scheme_id,n2kModel-methods}
\title{Get the scheme_id of a n2kModel object}
\usage{
get_scheme_id(x)

\S4method{get_scheme_id}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the scheme_id of the object
}
\description{
Get the scheme_id of a n2kModel object
}

