% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_seed.R
\docType{methods}
\name{get_seed}
\alias{get_seed}
\alias{get_seed,n2kModel-method}
\alias{get_seed,n2kModel-methods}
\title{Get the seed of a n2kModel object}
\usage{
get_seed(x)

\S4method{get_seed}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the seed of the object
}
\description{
Get the seed of a n2kModel object
}

