% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_session_info.R
\docType{methods}
\name{get_session_info}
\alias{get_session_info}
\alias{get_session_info,n2kModel-method}
\alias{get_session_info,n2kModel-methods}
\title{Get the session info of a n2kModel object}
\usage{
get_session_info(x)

\S4method{get_session_info}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the seed of the object
}
\description{
Get the session info of a n2kModel object
}

