% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sha1.R
\docType{methods}
\name{get_sha1,inla-method}
\alias{get_sha1,inla-method}
\title{Calculate SHA hash for inla object}
\usage{
\S4method{get_sha1}{inla}(x)
}
\arguments{
\item{x}{the inla object}
}
\description{
Calculate SHA hash for inla object
}

