% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sha1.R
\docType{methods}
\name{get_sha1,glmerMod-method}
\alias{get_sha1,glmerMod-method}
\title{Calculate SHA hash for glmerMod object}
\usage{
\S4method{get_sha1}{glmerMod}(x)
}
\arguments{
\item{x}{the glmerMod object}
}
\description{
Calculate SHA hash for glmerMod object
}

