% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_species_group_id.R
\docType{methods}
\name{get_species_group_id}
\alias{get_species_group_id}
\alias{get_species_group_id,n2kModel-method}
\alias{get_species_group_id,n2kModel-methods}
\title{Get the species group id of a n2kModel object}
\usage{
get_species_group_id(x)

\S4method{get_species_group_id}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the species group id of the object
}
\description{
Get the species group id of a n2kModel object
}

