% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_status_fingerprint.R
\docType{methods}
\name{get_status_fingerprint}
\alias{get_status_fingerprint}
\alias{get_status_fingerprint,n2kModel-method}
\alias{get_status_fingerprint,n2kModel-methods}
\title{Get the status fingerprint of a n2kModel object}
\usage{
get_status_fingerprint(x)

\S4method{get_status_fingerprint}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the status fingerprint of the object
}
\description{
Get the status fingerprint of a n2kModel object
}

