% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_weight.R
\docType{methods}
\name{get_weight}
\alias{get_weight}
\alias{get_weight,n2kGlmerPoisson-method}
\alias{get_weight,n2kGlmerPoisson-methods}
\title{Get the weight of a n2kModel object}
\usage{
get_weight(x)

\S4method{get_weight}{n2kGlmerPoisson}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the weight of the object
}
\description{
Get the weight of a n2kModel object
}

