% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_result.R
\name{import_result}
\alias{import_result}
\title{Import the results into the database}
\usage{
import_result(result, result.channel)
}
\arguments{
\item{result}{A n2kResult object (e.g. output of get_result())}

\item{result.channel}{An open ODBC connection to the database}
}
\value{
invisble NULL
}
\description{
Import the results into the database
}

