% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_inverse.R
\name{inla_inverse}
\alias{inla_inverse}
\title{Calculate the mean and 95\% confidence interval from the inverse of a
marginal}
\usage{
inla_inverse(marginal)
}
\arguments{
\item{marginal}{The INLA marginal}
}
\description{
This is useful for calculating the variance of a random effect when the
marginal gives the precision of the random effect.
}
