% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_a.R
\docType{methods}
\name{make_a}
\alias{make_a}
\alias{make_a,Spde-method}
\title{Convert an \code{Spde} object to a A object}
\usage{
make_a(object, data)

\S4method{make_a}{Spde}(object, data)
}
\arguments{
\item{object}{The \code{Spde} object}
}
\description{
Convert an \code{Spde} object to a A object
}
