% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_obsolete_dataset.R
\name{mark_obsolete_dataset}
\alias{mark_obsolete_dataset}
\title{Find and mark obsolete datasets}
\usage{
mark_obsolete_dataset(channel)
}
\arguments{
\item{channel}{An open ODBC channl}
}
\description{
A dataset is obsolete when a newer version is available
}

