% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_difference.R
\name{moving_difference}
\alias{moving_difference}
\title{Calculate coefficients for a moving difference}
\usage{
moving_difference(n_year, duration, first_year = 1)
}
\arguments{
\item{n_year}{Number of available years in the data.}

\item{duration}{Number of years in the moving window.
If the number of available years is less than this value, the trend will be
calculated for the available years.}

\item{first_year}{First year of the data.
Only used to name the rows.}
}
\value{
A matrix with moving average coefficients
One row for each window and and one column for each year in the data.
The format of the row names is
\verb{difference_\{window mid point start\}_\{window mid point end\}_\{window length\}}.
\code{difference_2001.5_2010.5_4} is the difference of the average for period for
the years 2009 to 2012 compared to the period from 2000 to 2003.
}
\description{
Calculate coefficients for a moving difference
}
\examples{
moving_difference(6, 2)
moving_difference(6, 2, 2000)
}
