% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_trend.R
\name{moving_trend}
\alias{moving_trend}
\title{Calculate moving trend coefficients}
\usage{
moving_trend(n_year, duration, first_year = 0)
}
\arguments{
\item{n_year}{Number of available years in the data.}

\item{duration}{Number of years in the moving window.
If the number of available years is less than this value, the trend will be
calculated for the available years.}

\item{first_year}{First year of the data.
Only used to name the rows.}
}
\value{
A matrix with the moving trend coefficients.
One row for each window and and one column for each year in the data.
The format of the row names is \verb{trend_\{window mid point\}_\{window length\}}.
\code{trend_2001.5_4} is the trend for the years 2000 to 2003.
}
\description{
Calculate moving trend coefficients
}
\examples{
moving_trend(5, 3)
moving_trend(5, 4, 2000)
}
