% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n2k_glmer_poisson.R
\docType{methods}
\name{n2k_glmer_poisson}
\alias{n2k_glmer_poisson}
\alias{n2k_glmer_poisson,data.frame,ANY-method}
\alias{n2k_glmer_poisson,my_lmer-methods}
\alias{n2k_glmer_poisson,n2kGlmerPoisson,glmerMod-method}
\alias{n2k_glmer_poisson,n2kGlmerPoisson-methods}
\title{Create a n2kGlmerPoisson object}
\usage{
n2k_glmer_poisson(data, ..., model.fit)

\S4method{n2k_glmer_poisson}{data.frame,ANY}(data, ..., model.fit)

\S4method{n2k_glmer_poisson}{n2kGlmerPoisson,glmerMod}(data, ..., model.fit)
}
\arguments{
\item{data}{a data.frame with the data to analyse}

\item{...}{other arguments}

\item{model.fit}{The fitted model}
}
\description{
Create a n2kGlmerPoisson object
}

