% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_hurdle_imputed.R
\docType{methods}
\name{n2k_hurdle_imputed}
\alias{n2k_hurdle_imputed}
\alias{n2k_hurdle_imputed,n2kInla-method}
\alias{n2k_hurdle_imputed,n2kHurdleImputed-methods}
\title{Create an \code{n2kHurdleImputed} object}
\usage{
n2k_hurdle_imputed(presence, count)

\S4method{n2k_hurdle_imputed}{n2kInla}(presence, count)
}
\arguments{
\item{presence}{the \code{n2kInla} object for the presence model.}

\item{count}{the \code{n2kInla} object for the count model.}
}
\description{
A new \code{n2kHurdleImputed} model.
}
