% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_inla_poisson.R
\docType{methods}
\name{n2k_inla_poisson}
\alias{n2k_inla_poisson}
\alias{n2k_inla_poisson,data.frame,ANY-method}
\alias{n2k_inla_poisson,n2kInlaPoisson-methods}
\alias{n2k_inla_poisson,n2kInlaPoisson,inla-method}
\title{Create a n2kInlaPoisson object}
\usage{
n2k_inla_poisson(data, ..., model.fit)

\S4method{n2k_inla_poisson}{data.frame,ANY}(data, status = "new",
  result.datasource.id, scheme.id, formula, species.group.id,
  location.group.id, model.type, first.imported.year, last.imported.year,
  duration, last.analysed.year, analysis.date, lin.comb = NULL,
  minimum = "", imputation.size, parent = character(0), seed,
  replicate.name = list(), parent.status = "converged",
  parent.statusfingerprint, ..., model.fit)

\S4method{n2k_inla_poisson}{n2kInlaPoisson,inla}(data, status,
  raw.imputed = NULL, ..., model.fit)
}
\arguments{
\item{data}{a data.frame with the data to analyse}

\item{...}{other arguments. See below}

\item{model.fit}{The fitted model}

\item{status}{a single character indicating the status of the model. Defaults to \code{"waiting"}.}

\item{result.datasource.id}{the id of the results datasource}

\item{scheme.id}{a single integer holding the id of the scheme.}

\item{formula}{a single character identifying the comparison}

\item{species.group.id}{a single integer identifing the species group}

\item{location.group.id}{a single integer identifing the location group}

\item{model.type}{the type of the models. Must start with \code{"inla comparison:"}}

\item{first.imported.year}{Oldest year considered in the data}

\item{last.imported.year}{Most recent year considered in the data}

\item{duration}{The width of the moving window. Defaults to the last.imported.year - first.imported.year + 1}

\item{last.analysed.year}{Most recent year in the window. Defaults to \code{last.imported.year}}

\item{analysis.date}{A POSIXct date indicating the date that the dataset was imported}

\item{lin.comb}{A model matrix to calculate linear combinations}

\item{imputation.size}{The required number of imputations defaults to 0.}

\item{seed}{a single integer used as a seed for all calculations. A random seed will be inserted when missing.}

\item{replicate.name}{A list with the names of replicates. Defaults to an empty list. Used in case of \code{f(X, ..., replicate = Z)}}. Should be a named list like e.g. \code{list(X = c("a", "b", "c")).}

\item{parent.status}{a \code{data.frame} with columns \code{ParentAnalysis} (the file fingerprint of the parent), \code{ParentStatusFingerprint} (the status fingerprint of the parent), and \code{ParentStatus} (the status of the parent)}
}
\description{
A new n2kInlaPoisson model is created when \code{data} is a data.frame.

In case \code{data} a n2kInlaPoisson object is, then only the model and status are updated. All other slots are unaffected.
}
