% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_manifest.R
\docType{methods}
\name{n2k_manifest}
\alias{n2k_manifest}
\alias{n2k_manifest,data.frame-method}
\alias{n2k_manifest,n2kManifest-methods}
\title{Create an \code{n2kManifest} object}
\usage{
n2k_manifest(manifest)

\S4method{n2k_manifest}{data.frame}(manifest)
}
\arguments{
\item{manifest}{a \code{data.frame} with the manifest content}
}
\description{
A new \code{n2kManifest} object is created when \code{manifest} is a
\code{data.frame}.
}
