% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_model_imputed.R
\docType{methods}
\name{n2k_model_imputed}
\alias{n2k_model_imputed}
\alias{n2k_model_imputed,ANY-method}
\alias{n2k_model_imputed,n2kModelImputed-methods}
\title{Create a n2kModelImputed object}
\usage{
n2k_model_imputed(...)

\S4method{n2k_model_imputed}{ANY}(...)
}
\arguments{
\item{...}{other arguments. See below}
}
\description{
Create a n2kModelImputed object

A new n2kModelImputed model.
}
\details{
\describe{
   \item{\code{scheme.id}}{a single integer holding the id of the scheme.}
   \item{\code{species.group.id}}{a single integer identifing the species group}
   \item{\code{location.group.id}}{a single integer identifing the location group}
   \item{\code{model.type}}{a single character identifying the type of model to fit to the data}
   \item{\code{formula}}{a single character holding the model formula}
   \item{\code{first.imported.year}}{Oldest year considered in the data}
   \item{\code{last.imported.year}}{Most recent year considered in the data}
   \item{\code{duration}}{The width of the moving window. Defaults to the last.imported.year - first.imported.year + 1}
   \item{\code{last.analysed.year}}{Most recent year in the window. Defaults to \code{last.imported.year}}
   \item{\code{analysis.date}}{A POSIXct date indicating the date that the dataset was imported}
   \item{\code{seed}}{a single integer used as a seed for all calculations. A random seed will be inserted when missing.}
   \item{\code{model.fun}}{an option list for the \code{model.fun} argument of \code{\link[multimput]{model_impute}}}
   \item{\code{model.args}}{an option list for the \code{model.args} argument of \code{\link[multimput]{model_impute}}}
   \item{\code{extractor}}{an option list for the \code{extractor} argument of \code{\link[multimput]{model_impute}}}
   \item{\code{extractor.args}}{an option list for the \code{extractor.args} argument of \code{\link[multimput]{model_impute}}}
   \item{\code{filter}}{an option list for the \code{filter} argument of \code{\link[multimput]{model_impute}}}
   \item{\code{mutate}}{an option list for the \code{mutate} argument of \code{\link[multimput]{model_impute}}}
  }
}

