% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_spde.R
\docType{methods}
\name{n2k_spde}
\alias{n2k_spde}
\alias{n2k_spde,data.frame,ANY-method}
\alias{n2k_spde,n2kSpde-methods}
\alias{n2k_spde,n2kSpde,inla-method}
\title{Create an \code{n2kSpde} object}
\usage{
n2k_spde(data, ..., model_fit)

\S4method{n2k_spde}{data.frame,ANY}(
  data,
  status = "new",
  result_datasource_id,
  scheme_id,
  family = "poisson",
  formula,
  species_group_id,
  location_group_id,
  model_type,
  spde,
  first_imported_year,
  last_imported_year,
  duration,
  last_analysed_year,
  analysis_date,
  lin_comb = NULL,
  minimum = "",
  imputation_size,
  parent = character(0),
  seed,
  replicate_name = list(),
  control = list(),
  parent_status = "converged",
  parent_statusfingerprint,
  extra,
  ...,
  model_fit
)

\S4method{n2k_spde}{n2kSpde,inla}(data, status, raw_imputed = NULL, ..., model_fit)
}
\arguments{
\item{data}{a \code{data.frame} with the data to analyse}

\item{...}{other arguments. See below}

\item{model_fit}{The fitted model}

\item{status}{A single character indicating the status of the model.
Defaults to \code{"waiting"}.}

\item{result_datasource_id}{A string identifying the data source.}

\item{scheme_id}{A single integer holding the id of the scheme.}

\item{family}{the family to use in the INLA model.}

\item{formula}{A single character identifying the comparison.}

\item{species_group_id}{A string identifying the species group.}

\item{location_group_id}{A string identifying the location group.}

\item{model_type}{The type of the models.
Must start with \code{"inla comparison:"}.}

\item{first_imported_year}{Oldest year considered in the data.}

\item{last_imported_year}{Most recent year considered in the data.}

\item{duration}{The width of the moving window.
Defaults to the \code{last_imported_year - first_imported_year + 1}.}

\item{last_analysed_year}{Most recent year in the window.
Defaults to \code{last_imported_year}.}

\item{analysis_date}{A \code{POSIXct} date indicating the date that the dataset
was imported.}

\item{lin_comb}{A model matrix to calculate linear combinations.}

\item{minimum}{The name of the variable which holds the minimum counts.
Only relevant in case of multiple imputation.}

\item{imputation_size}{The required number of imputations defaults to 0.}

\item{parent}{The file fingerprint of the optional parent analysis.}

\item{seed}{A single integer used as a seed for all calculations.
A random seed will be inserted when missing.}

\item{replicate_name}{A list with the names of replicates.
Defaults to an empty list.
Used in case of \code{f(X, ..., replicate = Z)}.
Should be a named list like e.g. \code{list(X = c("a", "b", "c"))}.}

\item{control}{A named list passed to \code{\link[INLA:inla]{INLA::inla()}} when fitting
the model.}

\item{parent_status}{A \code{data.frame} with columns
\code{parent_analysis} (the file fingerprint of the parent),
\code{parentstatus_fingerprint} (the status fingerprint of the parent),
and \code{parent_status} (the status of the parent).}

\item{parent_statusfingerprint}{The status fingerprint of the parent
analysis.}

\item{extra}{a \code{data.frame} with extra observations not used in the model.
They will be added in subsequent analyses.}

\item{raw_imputed}{the optional \code{rawImputed} object}
}
\description{
A new \code{n2kSpde} model is created when \code{data} is a \code{data.frame}.

In case \code{data} is an \code{n2kSpde} object, then only the model and
status are updated.
All other slots are unaffected.
}
