% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_aggregate_class.R
\docType{class}
\name{n2kAggregate-class}
\alias{n2kAggregate-class}
\title{The \code{n2kAggregate} class}
\description{
It holds analysis data based on an aggregated imputation
}
\section{Slots}{

\describe{
\item{\code{RawImputed}}{A \code{rawImputed} object with multiple
imputations.}
\item{\code{Function}}{The function to apply on each group.}
\item{\code{Filter}}{The object to pass to the \code{filter} argument of
\code{\link[multimput]{aggregate_impute}}.}
\item{\code{Join}}{The object to pass to the \code{join} argument of
\code{\link[multimput]{aggregate_impute}}.}
\item{\code{AggregatedImputed}}{An \code{aggregatedImputed} object with
multiple imputations.}
}
}

