% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n2kComposite_class.R
\docType{class}
\name{n2kComposite-class}
\alias{n2kComposite-class}
\title{The n2kComposite class}
\description{
Calculate composite indices from multiple analysis
}
\section{Slots}{

  \describe{
   \item{\code{Parent}}{a vector with the FileFingerprints of the parent analysis}
   \item{\code{ParentStatus}}{A data.frame with the FileFingerprint, StatusFingerprint and Status of the parent analysis}
   \item{\code{Parameter}}{A data.frame with the relevant parameter estimates of each parent analysis}
   \item{\code{Index}}{The composite index based on the parameters}
  }
}

