% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_hurdle_imputed_class.R
\docType{class}
\name{n2kHurdleImputed-class}
\alias{n2kHurdleImputed-class}
\title{The \code{n2kHurdleImputed} class}
\description{
It holds the model2 of hurdle imputed data.
}
\section{Slots}{

\describe{
\item{\code{Presence}}{A \code{rawImputed} object with the presence component.}

\item{\code{Count}}{A \code{rawImputed} object with the count component.}

\item{\code{Hurdle}}{A \code{aggregatedImputed} object with the resulting hurdle object.}
}}

