% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_import_class.R
\docType{class}
\name{n2kImport-class}
\alias{n2kImport-class}
\title{The \code{n2kImport} class}
\description{
It hold the imported datasets
}
\section{Slots}{

\describe{
\item{\code{Dataset}}{A \code{data.frame} with the datasets.}
}}

