% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2kInlaComparison_class.R
\docType{class}
\name{n2kInlaComparison-class}
\alias{n2kInlaComparison-class}
\title{The n2kInlaComparison class}
\description{
Compare multiple models using the WAIC criterion
}
\section{Slots}{

\describe{
\item{\code{WAIC}}{A \code{data.frame} with WAIC values per model.}
}
}

