% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2kInlaNbinomial_class.R
\docType{class}
\name{n2kInlaNbinomial-class}
\alias{n2kInlaNbinomial-class}
\title{The n2kInlaNBinomial class}
\description{
It hold analysis data based on a INLA negative binomial model
}
\section{Slots}{

\describe{
\item{\code{Data}}{a data.frame with the data}
\item{\code{LinearCombination}}{an optional matrix with the linear
combinations}
\item{\code{ReplicateName}}{an optional list with names of replicates}
\item{\code{Model}}{Either NULL or the resulting INLA model.}
\item{\code{ImputationSize}}{The number of multiple imputations. Defaults to 0, indication no multiple imputation.}
\item{\code{Minimum}}{an optional string containing the name of the variable in \code{Data} holding the minimal values for imputation}
\item{\code{RawImputed}}{A \code{rawImputed} object with multiple imputations.}
}
}

