% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n2kModel_class.R
\docType{class}
\name{n2kModel-class}
\alias{n2kModel-class}
\title{The n2kModel class}
\description{
A virtual superclass to contain the analysis models
}
\section{Slots}{

  \describe{
   \item{\code{Data}}{a data.frame with the data}
   \item{\code{Status}}{a single character indicating the status of the model}
   \item{\code{Seed}}{a single integer uses as a seed for all calculations}
  }
}

