% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_spde_class.R
\docType{class}
\name{n2kSpde-class}
\alias{n2kSpde-class}
\title{The \code{n2kSpde} class}
\description{
It hold analysis data based on an INLA model with SPDE.
}
\section{Slots}{

\describe{
\item{\code{Spde}}{A \code{list} containing the information for the SPDE.}
}}

