% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_yaml_to_bash.R
\name{order_manifest}
\alias{order_manifest}
\title{Convert a manifest in vector of analysis fingerprints}
\usage{
order_manifest(manifest)
}
\arguments{
\item{manifest}{the \code{n2kManifest}}
}
\description{
The order of the analysis is determined by the parent-child relationship.
It starts with the analyses without parents.
Then it adds the analyses with parents that have already been added.
This process is repeated until all analyses have been added.
}
