% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parent_status.R
\docType{methods}
\name{parent_status}
\alias{parent_status}
\alias{parent_status,n2kComposite-method}
\alias{parent_status,n2kComposite-methods}
\alias{parent_status,n2kLrtGlmer-method}
\alias{parent_status,n2kLrtGlmer-methods}
\title{Get the parent status of a n2kModel}
\usage{
parent_status(x)

\S4method{parent_status}{n2kLrtGlmer}(x)

\S4method{parent_status}{n2kComposite}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the parent status of the object
}
\description{
Get the parent status of a n2kModel
}

