% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_manifest.R
\docType{methods}
\name{read_manifest}
\alias{read_manifest}
\alias{read_manifest,character-method}
\alias{read_manifest,s3_bucket-method}
\title{Read a \code{n2kManifest} object}
\usage{
read_manifest(base, project, hash)

\S4method{read_manifest}{character}(base, project, hash)

\S4method{read_manifest}{s3_bucket}(base, project, hash)
}
\arguments{
\item{base}{The base location to read the manifest.}

\item{project}{Will be a relative path within the base location.}

\item{hash}{Optional the \code{sha1} of the manifest.
This can be abbreviated to to first unique characters.
The function will return an error in case of multiple matches.
If missing, then most recent manifest will be returned.}
}
\description{
Read a \code{n2kManifest} object
}
