% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_result.R
\docType{methods}
\name{read_result}
\alias{read_result}
\alias{read_result,ANY,character-method}
\alias{read_result,ANY,s3_bucket-method}
\alias{read_result,ANY,ANY-method}
\title{Read an \code{n2kResult} object}
\usage{
read_result(x, base, project)

\S4method{read_result}{ANY,character}(x, base, project)

\S4method{read_result}{ANY,s3_bucket}(x, base, project)

\S4method{read_result}{ANY,ANY}(x, base, project)
}
\arguments{
\item{x}{the file fingerprint of the \code{n2kResult}}

\item{base}{the base location to read the results}

\item{project}{will be a relative path within the base location}
}
\description{
Read an \code{n2kResult} object
}
