% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_estimate.R,
%   R/result_estimate_character.R, R/result_estimate_n2k_result.R
\docType{methods}
\name{result_estimate}
\alias{result_estimate}
\alias{result_estimate,character-method}
\alias{result_estimate,n2kResult-method}
\title{Aggregate all results estimates in a single dataframe}
\usage{
result_estimate(x, ...)

\S4method{result_estimate}{character}(x, ..., base, project)

\S4method{result_estimate}{n2kResult}(x, ...)
}
\arguments{
\item{x}{object with the current results}

\item{...}{further arguments (see Details)}

\item{base}{the base location to read the results}

\item{project}{will be a relative path within the base location}
}
\description{
Aggregate all results estimates in a single dataframe
}
