% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_metadata.R,
%   R/result_metadata_character.R, R/result_metadata_n2k_result.R
\docType{methods}
\name{result_metadata}
\alias{result_metadata}
\title{Aggregate all results meta data in a single dataframe}
\usage{
result_metadata(x, ...)

\S4method{result_metadata}{character}(x, ..., base, project)

\S4method{result_metadata}{n2kResult}(x, ...)
}
\arguments{
\item{x}{object with the current results}

\item{...}{further arguments (see Details)}

\item{base}{the base location to read the results}

\item{project}{will be a relative path within the base location}
}
\description{
Aggregate all results meta data in a single dataframe
}
