% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_single_model.R
\name{run_single_model}
\alias{run_single_model}
\title{Run a single element from the to do list}
\usage{
run_single_model(file)
}
\arguments{
\item{file}{The name of the rda file}
}
\description{
Run a single element from the to do list
}

