% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_factor_count_strictly_positive.R
\name{select_factor_count_strictly_positive}
\alias{select_factor_count_strictly_positive}
\title{Select data based on the number of prescences per category}
\usage{
select_factor_count_strictly_positive(observation, variable, threshold,
  relative = FALSE, dimension = 1)
}
\arguments{
\item{observation}{the \code{data.frame} with observations}

\item{variable}{the name of the \code{factor}}

\item{threshold}{the minimal threshold}

\item{relative}{When FALSE the threshold is the number of non-zero
observations.
When TRUE the threshold is the proportion of non-zero observations.
Defaults to FALSE.}

\item{dimension}{Indicates which element of \code{variable} is used for the
final aggregation.}
}
\description{
Prescences have \eqn{Count > 0}.
}
\examples{
observation <- data.frame(
  Count = c(4, 4, 4, 4, 3, 3, 3, 0, 2, 2, 0, 0),
  LocationID = rep(1:3, each = 4),
  Year = rep(c(1, 1, 1, 1, 2, 2), 2)
)
# Select the locations with at least 3 prescenses
select_factor_count_strictly_positive(
  observation,
  variable = "LocationID",
  threshold = 3
)
# Select those locations in which the species is present in at least 2 years
select_factor_count_strictly_positive(
  observation, variable = c("LocationID", "Year"), threshold = 2
)
# Select those years in which the species is present in at least 2 locations
select_factor_count_strictly_positive(
  observation, variable = c("LocationID", "Year"),
  threshold = 2,
  dimension = 2
)
}
