% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_factor_threshold.R
\name{select_factor_threshold}
\alias{select_factor_threshold}
\title{Select the observations based on the average of a factor}
\usage{
select_factor_threshold(observation, variable, threshold)
}
\arguments{
\item{observation}{the \code{data.frame} with observations}

\item{variable}{the name of the \code{factor}}

\item{threshold}{the minimal threshold}
}
\description{
The negative binomial average of the \code{Count} variable is calculated for each level of \code{variable}. Only the levels which are equal or larger than \code{threshold} times the maximal average (in the original scale) are retained.
}
\examples{
observation <- data.frame(
  Count = c(100, 101, 50, 51, 1, 0, 0, 0),
  LocationID = factor(rep(1:4, each = 2))
)
select_factor_threshold(observation, "LocationID", threshold = 0.05)
}

