% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_factor_treshold.R
\name{select_factor_treshold}
\alias{select_factor_treshold}
\title{Select the observations based on the average of a factor}
\usage{
select_factor_treshold(observation, variable, treshold)
}
\arguments{
\item{observation}{the \code{data.frame} with observations}

\item{variable}{the name of the \code{factor}}

\item{treshold}{the minimal treshold}
}
\description{
The negative binomial average of the \code{Count} variable is calculated for each level of \code{variable}. Only the levels which are equal or larger than \code{treshold} times the maximal average (in the original scale) are retained.
}
\examples{
observation <- data.frame(
  Count = c(100, 101, 50, 51, 1, 0, 0, 0),
  LocationID = factor(rep(1:4, each = 2))
)
select_factor_treshold(observation, "LocationID", treshold = 0.05)
}

