% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_observed_range.R
\name{select_observed_range}
\alias{select_observed_range}
\title{Limit the observation to the range in which the species is present}
\usage{
select_observed_range(observation, variable)
}
\arguments{
\item{observation}{the \code{data.frame} with observations}

\item{variable}{the name of the \code{factor}}
}
\description{
Limit the observation to the range in which the species is present
}
\examples{
observation <- data.frame(
  Count = c(0, 0, 100, 101, 0, 51, 1, 0, 0, 0),
  Year = 1:10
)
select_observed_range(observation, "Year")
}
