% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_package.R
\docType{methods}
\name{session_package}
\alias{session_package}
\alias{session_package,sessionInfo-method}
\alias{session_package,sessionInfo-methods}
\title{Convert a \code{sessionInfo()} to a data.frame of packages}
\usage{
session_package(session)

\S4method{session_package}{sessionInfo}(session)
}
\arguments{
\item{session}{The output of \code{sessionInfo()}}
}
\value{
a data.frame with the packages of a \code{sessionInfo()}
}
\description{
Convert a \code{sessionInfo()} to a data.frame of packages
}
