% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_result.R
\name{simplify_result}
\alias{simplify_result}
\title{simplify an n2kResult object}
\usage{
simplify_result(result)
}
\arguments{
\item{result}{a n2kResult object}
}
\description{
\itemize{
 \item Convert FileFingerprint to factor
 \item Convert all other fingerprints to integer
 \item Convert most of the characters to factors
}
}

