% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_class.R
\docType{class}
\name{Spde-class}
\alias{Spde-class}
\title{The \code{spde} class}
\description{
It holds the coordinates, range and sigma parameters for the SPDE model.
}
\section{Slots}{

\describe{
\item{\code{Coordinates}}{a \code{data.frame} with the coordinates used for the mesh.}

\item{\code{Range}}{a numeric vector of length 2.
Will be used as the \code{prior.range} argument of \verb{[INLA::inla.spde2.pcmatern]}.}

\item{\code{Sigma}}{a numeric vector of length 2.
Will be used as the \code{prior.sigma} argument of \verb{[INLA::inla.spde2.pcmatern]}.}
}}

