% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde2matern.R
\docType{methods}
\name{spde2matern}
\alias{spde2matern}
\title{Convert an \code{Spde} object to a matern object}
\usage{
spde2matern(object)
}
\arguments{
\item{object}{The \code{Spde} object}
}
\description{
Convert an \code{Spde} object to a matern object
}
