% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde2matern.R, R/spde2mesh.R
\docType{methods}
\name{spde2matern,Spde-method}
\alias{spde2matern,Spde-method}
\alias{spde2mesh}
\alias{spde2mesh,Spde-method}
\title{Convert an \code{Spde} object to a mesh object}
\usage{
\S4method{spde2matern}{Spde}(object)

spde2mesh(object)

\S4method{spde2mesh}{Spde}(object)
}
\arguments{
\item{object}{The \code{Spde} object}
}
\description{
Convert an \code{Spde} object to a mesh object
}
