% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\docType{methods}
\name{status<-}
\alias{status<-}
\alias{status<-,n2kGlmerPoisson-method}
\alias{status<-,n2kInlaNbinomial-method}
\alias{status<-,n2kInlaPoisson-method}
\alias{status<-,n2kLrtGlmer-method}
\alias{status<-,n2kComposite-method}
\alias{status<-,n2kInlaComparison-method}
\alias{status<-,n2kAggregate-method}
\alias{status<-,n2kModelImputed-method}
\title{Overwrite the status of a n2kAnalysisMetadata}
\usage{
status(x) <- value

\S4method{status}{n2kGlmerPoisson}(x) <- value

\S4method{status}{n2kInlaNbinomial}(x) <- value

\S4method{status}{n2kInlaPoisson}(x) <- value

\S4method{status}{n2kLrtGlmer}(x) <- value

\S4method{status}{n2kComposite}(x) <- value

\S4method{status}{n2kInlaComparison}(x) <- value

\S4method{status}{n2kAggregate}(x) <- value

\S4method{status}{n2kModelImputed}(x) <- value
}
\arguments{
\item{x}{the n2kAnalysisMetadata object}

\item{value}{the new values for the status}
}
\description{
Overwrite the status of a n2kAnalysisMetadata
}
