% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_manifest_yaml.R
\name{store_manifest_yaml}
\alias{store_manifest_yaml}
\title{Store to docker configuration on S3}
\usage{
store_manifest_yaml(x, base, project, docker, dependencies)
}
\arguments{
\item{x}{the n2kManifest}

\item{base}{the base location to store the manifest}

\item{project}{will be a relative path within the base location}

\item{docker}{the docker image to use}

\item{dependencies}{extra github package to install}
}
\description{
Store to docker configuration on S3
}
