% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{test_data}
\alias{test_data}
\title{A function to generate a simple dataset for unit testing}
\usage{
test_data(datasource.id = sha1(letters), missing = 0)
}
\arguments{
\item{datasource.id}{the string id of the datasource}

\item{missing}{the required proportion of missing data}
}
\description{
A function to generate a simple dataset for unit testing
}

