% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validOject.R
\docType{methods}
\name{validObject,character-method}
\alias{validObject,character-method}
\alias{validObject,character-methods}
\title{Get the validity of objects in files}
\usage{
\S4method{validObject}{character}(object, test = FALSE, complete = FALSE)
}
\arguments{
\item{object}{a single filename, a single path or a vector with filenames.}

\item{test}{See \code{\link[methods]{validObject}}}

\item{complete}{See \code{\link[methods]{validObject}}}
}
\description{
Get the validity of objects in files
}
