% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract}
\alias{extract}
\alias{extract,ANY,character-method}
\alias{extract,character-methods}
\alias{extract,ANY,n2kInlaNbinomial-method}
\alias{extract,n2kInlaNbinomial-methods}
\title{Extract the relevant coefficients}
\usage{
extract(extractor, object, path)

\S4method{extract}{ANY,character}(extractor, object, path)

\S4method{extract}{ANY,n2kInlaNbinomial}(extractor, object, path = NULL)
}
\arguments{
\item{extractor}{the extractor function}

\item{object}{the n2kModel object}

\item{path}{the optional path}
}
\value{
the relevant coefficients
}
\description{
Extract the relevant coefficients
}
