% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_every_model.R
\name{fit_every_model}
\alias{fit_every_model}
\title{Fit the model to the analysis files}
\usage{
fit_every_model(path = ".")
}
\arguments{
\item{path}{The path of the analysis files}
}
\description{
Fit the model to the analysis files
}

