% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_every_model.R
\name{fit_every_model}
\alias{fit_every_model}
\title{Fit the model to the analysis files}
\usage{
fit_every_model(path, status, verbose = TRUE, n.cluster = 1)
}
\arguments{
\item{path}{The path containing the analysis files.}

\item{status}{A vector with status levels to (re-)fit the model.
Defaults to \code{c("new", "waiting")} when missing.}

\item{verbose}{Show the name of the current analysis file on screen.
Defaults to \code{TRUE}}

\item{n.cluster}{The number of clusters to use.}
}
\description{
Fit the model to the analysis files
}
