% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\docType{methods}
\name{fit_model}
\alias{fit_model}
\alias{fit_model,character,character-method}
\alias{fit_model,n2kGlmerPoisson,character-method}
\alias{fit_model,n2kInlaNbinomial,character-method}
\title{Fit a n2kModel object}
\usage{
fit_model(x, status)

\S4method{fit_model}{character,character}(x, status)

\S4method{fit_model}{n2kGlmerPoisson,character}(x, status = c("new", "error",
  "converged", "false convergence"))

\S4method{fit_model}{n2kInlaNbinomial,character}(x, status = c("new", "error",
  "converged", "false convergence"))
}
\arguments{
\item{x}{the n2kModel}

\item{status}{A vector of statusses for which the model will be calculated}
}
\description{
Fit a n2kModel object
}

