% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R, R/fit_model_character.R,
%   R/fit_model_n2kAggregate.R, R/fit_model_n2kComposite.R,
%   R/fit_model_n2kGlmerPoisson.R, R/fit_model_n2kInla.R,
%   R/fit_model_n2kInlaComparison.R, R/fit_model_n2kLrtGlmer.R,
%   R/fit_model_n2kManifest.R, R/fit_model_n2kModelImputed.R,
%   R/fit_model_s3_object.R
\docType{methods}
\name{fit_model}
\alias{fit_model}
\alias{fit_model,character-method}
\alias{fit_model,n2kAggregate-method}
\alias{fit_model,n2kComposite-method}
\alias{fit_model,n2kGlmerPoisson-method}
\alias{fit_model,n2kInla-method}
\alias{fit_model,n2kInlaComparison-method}
\alias{fit_model,n2kLrtGlmer-method}
\alias{fit_model,n2kManifest-method}
\alias{fit_model,n2kModelImputed-method}
\alias{fit_model,s3_object-method}
\title{Fit a n2kModel object}
\usage{
fit_model(x, ...)

\S4method{fit_model}{character}(x, base, project, status = c("new",
  "waiting"), verbose = TRUE, ..., bucket)

\S4method{fit_model}{n2kAggregate}(x, ...)

\S4method{fit_model}{n2kComposite}(x, base, project, status = "new", ...)

\S4method{fit_model}{n2kGlmerPoisson}(x, ...)

\S4method{fit_model}{n2kInla}(x, status = "new", ..., timeout = NULL)

\S4method{fit_model}{n2kInlaComparison}(x, base, project, status = "new",
  verbose = TRUE, ...)

\S4method{fit_model}{n2kLrtGlmer}(x, ...)

\S4method{fit_model}{n2kManifest}(x, base, project, status = c("new",
  "waiting"), verbose = TRUE, ..., local, bash = FALSE)

\S4method{fit_model}{n2kModelImputed}(x, ...)

\S4method{fit_model}{s3_object}(x, ...)
}
\arguments{
\item{x}{the n2kModel}

\item{...}{other arguments. See details}

\item{base}{The root of a project. Can be either a directory on a file system
or an AWS S3 bucket object.
Extracted from \code{bucket} or \code{x} when missing.}

\item{project}{The subdirectory of the project. Is relative the \code{base}.
Extracted from \code{x} when missing.}

\item{status}{A vector with status levels naming the levels which should be
calculated.
Defaults to \code{"new"}.}

\item{verbose}{A logical indicating if the function should display the name
of the file and the status.
Defaults to \code{TRUE}.}

\item{bucket}{The name of the AWS S3 bucket.
Only used when \code{base} is missing.}

\item{timeout}{the optional number of second until the model will time out}

\item{local}{A local folder into which objects from an AWS S3 bucket are
downloaded.}

\item{bash}{Use the \code{littler} package do run the models in separate sessions.
This will release the memory.}
}
\description{
Fit a n2kModel object
}
\details{
\itemize{
\item \code{status}: A vector with status levels naming the levels which should be
recalculated.
Defaults to \code{"new"}.
\item \code{verbose}: A logical indicating if the function should display the name of
the file and the status.
Defaults to \code{TRUE}.
}
}
