% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R, R/fit_model_character.R,
%   R/fit_model_n2kAggregate.R, R/fit_model_n2kComposite.R,
%   R/fit_model_n2kGlmerPoisson.R, R/fit_model_n2kInlaComparison.R,
%   R/fit_model_n2kInlaNbinomial.R, R/fit_model_n2kLrtGlmer.R,
%   R/fit_model_n2kManifest.R, R/fit_model_n2kModelImputed.R,
%   R/fit_model_s3_object.R
\docType{methods}
\name{fit_model}
\alias{fit_model}
\alias{fit_model,character-method}
\alias{fit_model,n2kAggregate-method}
\alias{fit_model,n2kComposite-method}
\alias{fit_model,n2kGlmerPoisson-method}
\alias{fit_model,n2kInlaComparison-method}
\alias{fit_model,n2kInlaNbinomial-method}
\alias{fit_model,n2kLrtGlmer-method}
\alias{fit_model,n2kManifest-method}
\alias{fit_model,n2kModelImputed-method}
\alias{fit_model,s3_object-method}
\title{Fit a n2kModel object}
\usage{
fit_model(x, ...)

\S4method{fit_model}{character}(x, ...)

\S4method{fit_model}{n2kAggregate}(x, ...)

\S4method{fit_model}{n2kComposite}(x, ...)

\S4method{fit_model}{n2kGlmerPoisson}(x, ...)

\S4method{fit_model}{n2kInlaComparison}(x, ...)

\S4method{fit_model}{n2kInlaNbinomial}(x, ...)

\S4method{fit_model}{n2kLrtGlmer}(x, ...)

\S4method{fit_model}{n2kManifest}(x, ...)

\S4method{fit_model}{n2kModelImputed}(x, ...)

\S4method{fit_model}{s3_object}(x, ...)
}
\arguments{
\item{x}{the n2kModel}

\item{...}{other arguments. See details}
}
\description{
Fit a n2kModel object
}
\details{
\describe{
 \item{\code{status}}{A vector with status levels naming the levels which should be recalculated. Defaults to \code{"new"}}
 \item{\code{verbose}}{A logical indicating if the function should display the name of the file and the status. Defaults to \code{TRUE}}
}

\describe{
 \item{\code{status}}{A vector with status levels naming the levels which should be recalculated. Defaults to \code{"new"}}
 \item{\code{verbose}}{A logical indicating if the function should display the name of the file and the status. Defaults to \code{TRUE}}
}
}
