% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_model.R
\docType{methods}
\name{fit_model}
\alias{fit_model}
\alias{fit_model,character-method}
\alias{fit_model,n2kComposite-method}
\alias{fit_model,n2kGlmerPoisson-method}
\alias{fit_model,n2kInlaComparison-method}
\alias{fit_model,n2kInlaNbinomial-method}
\alias{fit_model,n2kLrtGlmer-method}
\title{Fit a n2kModel object}
\usage{
fit_model(x, ...)

\S4method{fit_model}{character}(x, ...)

\S4method{fit_model}{n2kGlmerPoisson}(x, ...)

\S4method{fit_model}{n2kInlaNbinomial}(x, ...)

\S4method{fit_model}{n2kLrtGlmer}(x, ...)

\S4method{fit_model}{n2kComposite}(x, ...)

\S4method{fit_model}{n2kInlaComparison}(x, ...)
}
\arguments{
\item{x}{the n2kModel}

\item{...}{other arguments. See details}
}
\description{
Fit a n2kModel object
}
\details{
\describe{
 \item{\code{status}}{A vector with status levels naming the levels which should be recalculated. Defaults to \code{"new"}}
 \item{\code{verbose}}{A logical indicating if the function should display the name of the file and the status. Defaults to \code{TRUE}}
}
}

