% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_analysis_version.R
\docType{methods}
\name{get_analysis_version}
\alias{get_analysis_version}
\alias{get_analysis_version,data.frame-method}
\alias{get_analysis_version,n2kModel-method}
\alias{get_analysis_version,n2kModel-methods}
\alias{get_analysis_version,sessionInfo-method}
\alias{get_analysis_version,sessionInfo-methods}
\title{Get an analysis version}
\usage{
get_analysis_version(version)

\S4method{get_analysis_version}{data.frame}(version)

\S4method{get_analysis_version}{n2kModel}(version)

\S4method{get_analysis_version}{sessionInfo}(version)
}
\arguments{
\item{version}{the object to extract the version}
}
\description{
Get an analysis version
}

