% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anomaly.R, R/get_anomaly_n2k_inla.R,
%   R/get_anomaly_n2k_model.R
\docType{methods}
\name{get_anomaly}
\alias{get_anomaly}
\alias{get_anomaly,n2kInla-method}
\alias{get_anomaly,n2kInla-methods}
\alias{get_anomaly,n2kModel-method}
\title{Get the anomalies from a model}
\usage{
get_anomaly(analysis, ...)

\S4method{get_anomaly}{n2kInla}(
  analysis,
  n = 20,
  expected_ratio = 5,
  expected_absent = 5,
  random_threshold = 1.05,
  verbose = TRUE,
  ...
)

\S4method{get_anomaly}{n2kModel}(analysis, verbose = TRUE, ...)
}
\arguments{
\item{analysis}{The model to add}

\item{...}{Extra options. See details.}

\item{n}{the number of anomalies per category.}

\item{expected_ratio}{Observations that have
\code{observed / fitted > expected_ratio} or \code{fitted / observed > expected_ratio}
are potential anomalies.
Defaults to \code{5}, which implies that observed values that are 5 times higher
or lower than the fitted values are potential anomalies.}

\item{expected_absent}{Zero observations where \code{fitted > expected_absent} are
potential anomalies.}

\item{random_threshold}{The minimal relative effect size of a random effect.
Random effect with a smaller effect size will never be an anomaly.
Defaults to 1.05 (5\%).}

\item{verbose}{Print extra information on the screen}
}
\description{
Get the anomalies from a model
}
