% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\docType{methods}
\name{get_data}
\alias{get_data}
\alias{get_data,n2kInla-method}
\alias{get_data,n2kInla-methods}
\title{Get the data of a n2kModel object}
\usage{
get_data(x)

\S4method{get_data}{n2kInla}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the data of the object
}
\description{
Get the data of a n2kModel object
}
