% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\docType{methods}
\name{get_data}
\alias{get_data}
\alias{get_data,n2kGlmerPoisson-method}
\alias{get_data,n2kGlmerPoisson-methods}
\alias{get_data,n2kInlaNbinomial-method}
\alias{get_data,n2kInlaNbinomial-methods}
\title{Get the data of a n2kModel object}
\usage{
get_data(x)

\S4method{get_data}{n2kGlmerPoisson}(x)

\S4method{get_data}{n2kInlaNbinomial}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the data of the object
}
\description{
Get the data of a n2kModel object
}

