% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_fingerprint.R
\docType{methods}
\name{get_file_fingerprint}
\alias{get_file_fingerprint}
\alias{get_file_fingerprint,n2kAnalysisMetadata-method}
\alias{get_file_fingerprint,n2kAnalysisMetadata-methods}
\title{Get the file fingerprint of a n2kModel object}
\usage{
get_file_fingerprint(x)

\S4method{get_file_fingerprint}{n2kAnalysisMetadata}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the file fingerprint of the object
}
\description{
Get the file fingerprint of a n2kModel object
}

