% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_location_group_id.R
\docType{methods}
\name{get_location_group_id}
\alias{get_location_group_id}
\alias{get_location_group_id,n2kAnalysisMetadata-method}
\alias{get_location_group_id,n2kAnalysisMetadata-methods}
\title{Get the location group id of a n2kModel object}
\usage{
get_location_group_id(x)

\S4method{get_location_group_id}{n2kAnalysisMetadata}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the location group id of the object
}
\description{
Get the location group id of a n2kModel object
}

