% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model.R
\docType{methods}
\name{get_model}
\alias{get_model}
\alias{get_model,n2kGlmerPoisson-method}
\alias{get_model,n2kGlmerPoisson-methods}
\alias{get_model,n2kInlaNbinomial-method}
\alias{get_model,n2kInlaNbinomial-methods}
\title{Get the model of a n2kModel object}
\usage{
get_model(x)

\S4method{get_model}{n2kGlmerPoisson}(x)

\S4method{get_model}{n2kInlaNbinomial}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the model of the object
}
\description{
Get the model of a n2kModel object
}

