% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_parameter.R,
%   R/get_model_parameter_n2kAggregated.R, R/get_model_parameter_n2kComposite.R,
%   R/get_model_parameter_n2kInla.R, R/get_model_parameter_n2kInlaComparison.R,
%   R/get_model_parameter_n2kModelImputed.R
\docType{methods}
\name{get_model_parameter}
\alias{get_model_parameter}
\alias{get_model_parameter,n2kAggregate-method}
\alias{get_model_parameter,n2kComposite-method}
\alias{get_model_parameter,n2kInla-method}
\alias{get_model_parameter,n2kInlaComparison-method}
\alias{get_model_parameter,n2kModelImputed-method}
\title{Add the model parameters from a model}
\usage{
get_model_parameter(analysis, ...)

\S4method{get_model_parameter}{n2kAggregate}(analysis, ...)

\S4method{get_model_parameter}{n2kComposite}(analysis, ...)

\S4method{get_model_parameter}{n2kInla}(analysis, verbose = TRUE, ...)

\S4method{get_model_parameter}{n2kInlaComparison}(analysis, ...)

\S4method{get_model_parameter}{n2kModelImputed}(analysis, ...)
}
\arguments{
\item{analysis}{The model to add.}

\item{...}{extra options}
}
\description{
Add the model parameters from a model
}
