% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model_parameter.R
\docType{methods}
\name{get_model_parameter}
\alias{get_model_parameter}
\alias{get_model_parameter,n2kComposite-method}
\alias{get_model_parameter,n2kComposite-methods}
\alias{get_model_parameter,n2kGlmerPoisson-method}
\alias{get_model_parameter,n2kGlmerPoisson-methods}
\alias{get_model_parameter,n2kLrtGlmer-method}
\alias{get_model_parameter,n2kLrtGlmer-methods}
\title{Add the model parameters from a model}
\usage{
get_model_parameter(analysis)

\S4method{get_model_parameter}{n2kGlmerPoisson}(analysis)

\S4method{get_model_parameter}{n2kLrtGlmer}(analysis)

\S4method{get_model_parameter}{n2kComposite}(analysis)
}
\arguments{
\item{analysis}{The model to add}
}
\description{
Add the model parameters from a model
}

