% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_parameter.R,
%   R/get_model_parameter_n2kAggregated.R, R/get_model_parameter_n2kComposite.R,
%   R/get_model_parameter_n2kGlmerPoisson.R,
%   R/get_model_parameter_n2kInlaComparison.R,
%   R/get_model_parameter_n2kInlaNbinomial.R,
%   R/get_model_parameter_n2kLrtGlmer.R, R/get_model_parameter_n2kModelImputed.R
\docType{methods}
\name{get_model_parameter}
\alias{get_model_parameter}
\alias{get_model_parameter,n2kAggregate-method}
\alias{get_model_parameter,n2kComposite-method}
\alias{get_model_parameter,n2kGlmerPoisson-method}
\alias{get_model_parameter,n2kInlaComparison-method}
\alias{get_model_parameter,n2kInlaNbinomial-method}
\alias{get_model_parameter,n2kLrtGlmer-method}
\alias{get_model_parameter,n2kModelImputed-method}
\title{Add the model parameters from a model}
\usage{
get_model_parameter(analysis, ...)

\S4method{get_model_parameter}{n2kAggregate}(analysis, ...)

\S4method{get_model_parameter}{n2kComposite}(analysis, ...)

\S4method{get_model_parameter}{n2kGlmerPoisson}(analysis, verbose = TRUE,
  ...)

\S4method{get_model_parameter}{n2kInlaComparison}(analysis, ...)

\S4method{get_model_parameter}{n2kInlaNbinomial}(analysis,
  verbose = TRUE, ...)

\S4method{get_model_parameter}{n2kLrtGlmer}(analysis, ...)

\S4method{get_model_parameter}{n2kModelImputed}(analysis, ...)
}
\arguments{
\item{analysis}{The model to add}

\item{...}{extra options}

\item{verbose}{Print extra information on the screen}
}
\description{
Add the model parameters from a model
}
