% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model_type.R
\docType{methods}
\name{get_model_type}
\alias{get_model_type}
\alias{get_model_type,n2kModel-method}
\alias{get_model_type,n2kModel-methods}
\title{Get the model type of a n2kModel object}
\usage{
get_model_type(x)

\S4method{get_model_type}{n2kModel}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the model type of the object
}
\description{
Get the model type of a n2kModel object
}

