% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent.R
\docType{methods}
\name{get_parent}
\alias{get_parent}
\alias{get_parent,n2kLrtGlmer-method}
\alias{get_parent,n2kLrtGlmer-methods}
\alias{get_parent,n2kComposite-method}
\alias{get_parent,n2kComposite-methods}
\title{Get the fingerprint of the parent analysis of a n2kModel object}
\usage{
get_parent(x, ...)

\S4method{get_parent}{n2kLrtGlmer}(x, which.parent = c(0, 1))

\S4method{get_parent}{n2kComposite}(x, ...)
}
\arguments{
\item{x}{the n2kModel object}

\item{...}{additional arguments (see details)}

\item{which.parent}{0 = the parent for the null hypothesis, 1 =  the parent for the alternative hypothesis}
}
\value{
the parent of the object
}
\description{
Get the fingerprint of the parent analysis of a n2kModel object
}
