% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_result.R
\docType{methods}
\name{get_result}
\alias{get_result}
\alias{get_result,character-method}
\alias{get_result,n2kModel-method}
\alias{get_result,n2kModel-methods}
\title{Add the results from an analysis}
\usage{
get_result(x, ...)

\S4method{get_result}{n2kModel}(x, ...)

\S4method{get_result}{character}(x, ...)
}
\arguments{
\item{x}{object with the current results}

\item{...}{further arguments (see Details)}
}
\description{
Add the results from an analysis
}

