% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_result.R
\docType{methods}
\name{get_result}
\alias{get_result}
\alias{get_result,character-method}
\alias{get_result,n2kModel-method}
\alias{get_result,n2kModel-methods}
\title{Add the results from an analysis}
\usage{
get_result(x, ...)

\S4method{get_result}{n2kModel}(x, ...)

\S4method{get_result}{character}(x, keep.fingerprint = TRUE, n.cluster = 1,
  ...)
}
\arguments{
\item{x}{object with the current results}

\item{...}{further arguments (see Details)}

\item{keep.fingerprint}{Keep the character fingerprints? Otherwise change them into integers}

\item{n.cluster}{the number of clusters to run this function in parallel. Defaults to 1 (= no parallel computing).}
}
\description{
Add the results from an analysis
}

