% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R, R/get_result_character.R,
%   R/get_result_n2kInla.R, R/get_result_n2kModel.R, R/get_result_s3_object.R
\docType{methods}
\name{get_result}
\alias{get_result}
\alias{get_result,character-method}
\alias{get_result,n2kInla-method}
\alias{get_result,n2kModel-method}
\alias{get_result,s3_object-method}
\title{Add the results from an analysis}
\usage{
get_result(x, ...)

\S4method{get_result}{character}(x, n.cluster = 1, verbose = TRUE, ...)

\S4method{get_result}{n2kInla}(x, verbose = TRUE, ...)

\S4method{get_result}{n2kModel}(x, verbose = TRUE, ...)

\S4method{get_result}{s3_object}(x, ...)
}
\arguments{
\item{x}{object with the current results}

\item{...}{further arguments (see Details)}

\item{n.cluster}{the number of clusters to run this function in parallel. Defaults to 1 (= no parallel computing).}

\item{verbose}{Print extra information on the screen}
}
\description{
Add the results from an analysis
}
