% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scheme_id.R
\docType{methods}
\name{get_scheme_id}
\alias{get_scheme_id}
\alias{get_scheme_id,n2kAnalysisMetadata-method}
\alias{get_scheme_id,n2kAnalysisMetadata-methods}
\title{Get the \code{scheme_id} of an \code{n2kModel} object}
\usage{
get_scheme_id(x)

\S4method{get_scheme_id}{n2kAnalysisMetadata}(x)
}
\arguments{
\item{x}{the \code{n2kModel} object}
}
\value{
the \code{scheme_id} of the object
}
\description{
Get the \code{scheme_id} of an \code{n2kModel} object
}
