% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_group_id.R
\docType{methods}
\name{get_species_group_id}
\alias{get_species_group_id}
\alias{get_species_group_id,n2kAnalysisMetadata-method}
\alias{get_species_group_id,n2kAnalysisMetadata-methods}
\title{Get the species group id of an \code{n2kModel} object}
\usage{
get_species_group_id(x)

\S4method{get_species_group_id}{n2kAnalysisMetadata}(x)
}
\arguments{
\item{x}{the \code{n2kModel} object}
}
\value{
the species group id of the object
}
\description{
Get the species group id of an \code{n2kModel} object
}
