% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_yaml_to_bash.R
\docType{methods}
\name{manifest_yaml_to_bash}
\alias{manifest_yaml_to_bash}
\alias{manifest_yaml_to_bash,s3_bucket-method}
\alias{manifest_yaml_to_bash,character-method}
\title{Convert a manifest yaml file into a bash script}
\usage{
manifest_yaml_to_bash(base, project, hash)

\S4method{manifest_yaml_to_bash}{s3_bucket}(base, project, hash)

\S4method{manifest_yaml_to_bash}{character}(base, project, hash)
}
\arguments{
\item{base}{the base location to store the manifest}

\item{project}{will be a relative path within the base location}

\item{hash}{Fingerprint of the manifest \code{yaml}file.}
}
\description{
Convert a manifest yaml file into a bash script
}
