% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest_yaml_to_bash.R
\docType{methods}
\name{manifest_yaml_to_bash}
\alias{manifest_yaml_to_bash}
\alias{manifest_yaml_to_bash,s3_bucket-method}
\alias{manifest_yaml_to_bash,character-method}
\title{Convert a manifest yaml file into a bash script}
\usage{
manifest_yaml_to_bash(
  base,
  project,
  hash,
  shutdown = FALSE,
  split = 1,
  status = c("new", "waiting")
)

\S4method{manifest_yaml_to_bash}{s3_bucket}(
  base,
  project,
  hash,
  shutdown = FALSE,
  split = 1,
  status = c("new", "waiting")
)

\S4method{manifest_yaml_to_bash}{character}(
  base,
  project,
  hash,
  shutdown = FALSE,
  split = 1,
  status = c("new", "waiting")
)
}
\arguments{
\item{base}{the base location to store the manifest}

\item{project}{will be a relative path within the base location}

\item{hash}{Fingerprint of the manifest \code{yaml}file.}

\item{shutdown}{Append a shutdown command at the end of the script.
Defaults to \code{FALSE}.}

\item{split}{Number of scripts over which to splits the analyses.
Default to 1.}

\item{status}{A vector with status levels naming the levels which should be
calculated.
Defaults to \code{c("new", "waiting")}.}
}
\description{
Convert a manifest yaml file into a bash script
}
