% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n2k_glmer_poisson.R
\docType{methods}
\name{n2k_glmer_poisson}
\alias{n2k_glmer_poisson}
\alias{n2k_glmer_poisson,data.frame,ANY-method}
\alias{n2k_glmer_poisson,my_lmer-methods}
\alias{n2k_glmer_poisson,n2kGlmerPoisson,glmerMod-method}
\alias{n2k_glmer_poisson,n2kGlmerPoisson-methods}
\title{Create a n2kGlmerPoisson object}
\usage{
n2k_glmer_poisson(data, ..., model.fit)

\S4method{n2k_glmer_poisson}{data.frame,ANY}(data, ..., model.fit)

\S4method{n2k_glmer_poisson}{n2kGlmerPoisson,glmerMod}(data, ..., model.fit)
}
\arguments{
\item{data}{a data.frame with the data to analyse}

\item{...}{other arguments. See below}

\item{model.fit}{The fitted model}
}
\description{
Create a n2kGlmerPoisson object

A new n2kGlmerPoisson model is created when \code{data} is a data.frame.

In case \code{data} a n2kGlmerPoisson object is, then only the model and status are updated. All other slots are unaffected.
}
\details{
\describe{
   \item{\code{status}}{a single character indicating the status of the model. Defaults to 'new' when \code{data} is a data.frame.}
   \item{\code{scheme.id}}{a single integer holding the id of the scheme.}
   \item{\code{species.group.id}}{a single integer identifing the species group}
   \item{\code{location.group.id}}{a single integer identifing the location group}
   \item{\code{model.type}}{a single character identifying the type of model to fit to the data}
   \item{\code{covariate}}{a single character holding the right hand side of the model formula}
   \item{\code{first.imported.year}}{Oldest year considered in the data}
   \item{\code{last.imported.year}}{Most recent year considered in the data}
   \item{\code{duration}}{The width of the moving window. Defaults to the last.imported.year - first.imported.year + 1}
   \item{\code{last.analysed.year}}{Most recent year in the window. Defaults to \code{last.imported.year}}
   \item{\code{analysis.date}}{A POSIXct date indicating the date that the dataset was imported}
   \item{\code{weight}}{The name of the variable to use as weights. '' indicates no weighting. Defaults to ''}
   \item{\code{seed}}{a single integer used as a seed for all calculations. A random seed will be inserted when missing.}
  }
}

