% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_import.R
\docType{methods}
\name{n2k_import}
\alias{n2k_import}
\alias{n2k_import,ANY-method}
\title{Create an \code{n2kImport} object}
\usage{
n2k_import(...)

\S4method{n2k_import}{ANY}(...)
}
\arguments{
\item{...}{other arguments. See below}
}
\description{
A new \code{n2kImport} model.
}
\details{
\itemize{
\item \code{scheme_id}: a string holding the id of the scheme.
\item \code{species_group_id}: a string identifying the species group.
\item \code{location_group_id}: a string identifying the location group.
\item \code{model_type}: a string identifying the type of model to fit to the data.
\item \code{first_imported_year}: Oldest year considered in the data.
\item \code{last_imported_year}: Most recent year considered in the data.
\item \code{duration}: The width of the moving window.
Defaults to the \code{last_imported_year - first_imported_year + 1}.
\item \code{last_analysed_year}: Most recent year in the window.
Defaults to \code{last_imported_year}.
\item \code{analysis_date}: A \code{POSIXct} date indicating the date that the dataset was
imported.
\item \code{seed}: A single integer used as a seed for all calculations.
A random seed will be inserted when missing.
}

\itemize{
\item \code{formula}: a string holding the model formula.
\item \code{dataset}: A \code{data.frame} with \code{filename}, \code{fingerprint} and \code{import_date}.
}
}
