% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_inla_comparison.R
\docType{methods}
\name{n2k_inla_comparison}
\alias{n2k_inla_comparison}
\alias{n2k_inla_comparison,data.frame-method}
\alias{n2k_inla_comparison,n2kInlaComparison-methods}
\title{Create a n2kInlaComparison object}
\usage{
n2k_inla_comparison(parent_status, ...)

\S4method{n2k_inla_comparison}{data.frame}(
  parent_status,
  status = "waiting",
  result_datasource_id,
  scheme_id,
  formula,
  species_group_id,
  location_group_id,
  model_type,
  first_imported_year,
  last_imported_year,
  duration,
  last_analysed_year,
  analysis_date,
  ...,
  seed
)
}
\arguments{
\item{parent_status}{A \code{data.frame} with columns
\code{parent_analysis} (the file fingerprint of the parent),
\code{parentstatus_fingerprint} (the status fingerprint of the parent),
and \code{parent_status} (the status of the parent).}

\item{...}{other arguments}

\item{status}{A single character indicating the status of the model.
Defaults to \code{"waiting"}.}

\item{result_datasource_id}{A string identifying the datasource.}

\item{scheme_id}{A single integer holding the id of the scheme.}

\item{formula}{A single character identifying the comparison.}

\item{species_group_id}{A string identifying the species group.}

\item{location_group_id}{A string identifying the location group.}

\item{model_type}{The type of the models.
Must start with \code{"inla comparison:"}.}

\item{first_imported_year}{Oldest year considered in the data.}

\item{last_imported_year}{Most recent year considered in the data.}

\item{duration}{The width of the moving window.
Defaults to the \code{last_imported_year - first_imported_year + 1}.}

\item{last_analysed_year}{Most recent year in the window.
Defaults to \code{last_imported_year}.}

\item{analysis_date}{A POSIXct date indicating the date that the dataset was
imported.}

\item{seed}{A single integer used as a seed for all calculations.
A random seed will be inserted when missing.}
}
\description{
A new \code{n2kInlaComparison} model is created when \code{parent} is a
\code{character}.
}
