% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2k_composite_class.R
\docType{class}
\name{n2kComposite-class}
\alias{n2kComposite-class}
\title{The \code{n2kComposite} class}
\description{
Calculate composite indices from multiple analysis
}
\section{Slots}{

\describe{
\item{\code{Extractor}}{A function to extract the relevant parameters from
the model.}
\item{\code{Parameter}}{A data.frame with the relevant parameter estimates
of each parent analysis.}
\item{\code{Index}}{The composite index based on the parameters}
}
}

