% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n2kGlmerPoisson_class.R
\docType{class}
\name{n2kGlmerPoisson-class}
\alias{n2kGlmerPoisson-class}
\title{The n2kGlmerPoisson class}
\description{
It hold analysis data based on a glmer poisson model
}
\section{Slots}{

  \describe{
   \item{\code{Weight}}{The name of the variable to use as weights. '' indicates no weighting.}
   \item{\code{Model}}{Either NULL or the resulting glmer model.}
  }
}

