% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/n2kInlaNbinomial_class.R
\docType{class}
\name{n2kInlaNbinomial-class}
\alias{n2kInlaNbinomial-class}
\title{The n2kInlaNBinomial class}
\description{
It hold analysis data based on a INLA negative binomial model
}
\section{Slots}{

  \describe{
   \item{\code{Model}}{Either NULL or the resulting INLA model.}
  }
}

