% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2kLrtGlmer_class.R
\docType{class}
\name{n2kLrtGlmer-class}
\alias{n2kLrtGlmer-class}
\title{The n2kLrtGlmer class}
\description{
Calculate composite indices from multiple analysis
}
\section{Slots}{

\describe{
\item{\code{Parent0}}{The file fingerprint of the NULL model.}
\item{\code{Model}}{The alternative model.}
\item{\code{Model0}}{The NULL model.}
\item{\code{Anova}}{The anova table.}
}
}

