% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_status.R
\docType{methods}
\name{parent_status<-}
\alias{parent_status<-}
\alias{parent_status<-,n2kComposite-method}
\title{Overwrite the status of a n2kAnalysisMetadata}
\usage{
parent_status(x) <- value

\S4method{parent_status}{n2kComposite}(x) <- value
}
\arguments{
\item{x}{the n2kAnalysisMetadata object}

\item{value}{the new values for the status}
}
\description{
Overwrite the status of a n2kAnalysisMetadata
}
