% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha1.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sha1}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{digest}{\code{\link[digest]{sha1}}}
}}

