% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde.R
\name{spde}
\alias{spde}
\title{Create an \code{Spde} object}
\usage{
spde(coordinates, range, sigma)
}
\arguments{
\item{coordinates}{a \code{data.frame} of coordinates use to define the mesh.}

\item{range}{a numeric vector of length 2.
Will be used as the \code{prior.range} argument of \verb{[INLA::inla.spde2.pcmatern]}.}

\item{sigma}{a numeric vector of length 2.
Will be used as the \code{prior.sigma} argument of \verb{[INLA::inla.spde2.pcmatern]}.}
}
\description{
Create an \code{Spde} object
}
