% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{methods}
\name{status<-}
\alias{status<-}
\alias{status<-,n2kModel-method}
\title{Overwrite the status of a n2kModel}
\usage{
status(x) <- value

\S4method{status}{n2kModel}(x) <- value
}
\arguments{
\item{x}{the n2kModel object}

\item{value}{the new values for the status}
}
\description{
Overwrite the status of a n2kModel
}

