% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{methods}
\name{status<-}
\alias{status<-}
\alias{status<-,n2kComposite-method}
\alias{status<-,n2kGlmerPoisson-method}
\alias{status<-,n2kInlaComparison-method}
\alias{status<-,n2kInlaNbinomial-method}
\alias{status<-,n2kLrtGlmer-method}
\title{Overwrite the status of a n2kAnalysisMetadata}
\usage{
status(x) <- value

\S4method{status}{n2kGlmerPoisson}(x) <- value

\S4method{status}{n2kInlaNbinomial}(x) <- value

\S4method{status}{n2kLrtGlmer}(x) <- value

\S4method{status}{n2kComposite}(x) <- value

\S4method{status}{n2kInlaComparison}(x) <- value
}
\arguments{
\item{x}{the n2kAnalysisMetadata object}

\item{value}{the new values for the status}
}
\description{
Overwrite the status of a n2kAnalysisMetadata
}

