% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/status.R
\docType{methods}
\name{status}
\alias{status}
\alias{status,character-method}
\alias{status,n2kModel-method}
\alias{status,n2kModel-methods}
\title{Get the status of a n2kModel}
\usage{
status(x)

\S4method{status}{n2kModel}(x)

\S4method{status}{character}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the status of the object
}
\description{
Get the status of a n2kModel
}

