% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\docType{methods}
\name{status}
\alias{status}
\alias{status,character-method}
\alias{status,character-methods}
\alias{status,n2kAnalysisMetadata-method}
\alias{status,n2kAnalysisMetadata-methods}
\title{Get the status of a n2kModel}
\usage{
status(x)

\S4method{status}{n2kAnalysisMetadata}(x)

\S4method{status}{character}(x)
}
\arguments{
\item{x}{the n2kModel object}
}
\value{
the status of the object
}
\description{
Get the status of a n2kModel
}

