% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_manifest_yaml.R
\docType{methods}
\name{store_manifest_yaml}
\alias{store_manifest_yaml}
\alias{store_manifest_yaml,ANY,s3_bucket-method}
\alias{store_manifest_yaml,ANY,character-method}
\title{Store a Docker configuration}
\usage{
store_manifest_yaml(x, base, project, docker, dependencies)

\S4method{store_manifest_yaml}{ANY,s3_bucket}(x, base, project, docker, dependencies)

\S4method{store_manifest_yaml}{ANY,character}(x, base, project, docker, dependencies)
}
\arguments{
\item{x}{the \code{n2kManifest}}

\item{base}{the base location to store the manifest}

\item{project}{will be a relative path within the base location}

\item{docker}{the docker image to use}

\item{dependencies}{extra GitHub packages to install}
}
\description{
Store a Docker configuration
}
