% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_model.R
\docType{methods}
\name{store_model}
\alias{store_model}
\alias{store_model,ANY,character-method}
\alias{store_model,ANY,s3_bucket-method}
\title{Store a n2kModel object}
\usage{
store_model(x, base, project)

\S4method{store_model}{ANY,character}(x, base, project)

\S4method{store_model}{ANY,s3_bucket}(x, base, project)
}
\arguments{
\item{x}{the n2kModel}

\item{base}{the base location to store the model}

\item{project}{will be a relative path within the base location}
}
\description{
Store a n2kModel object
}
